+++ CHANGELOG.md
@@ -8,9 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- GPX reader demo
+- A bunch of streamlit snippets with real image processing examples
+- streamlit-webrtc library to handle video streaming
+
 ### Fixed
 
 - Incorrect indentation
+
+### Removed
+
+- Failure.png resource removed
+
 ## [1.5.1] - 2023-07-21
 
 ### Fixed